Add AssetFile for file helpers
