+++ assets/AssetBundle.h
@@ -19,37 +19,11 @@
 #include <unordered_map>
 #include <vector>
 
-#include "assets/Asset.h"
 #include "assets/AssetLump.h"
+#include "assets/AssetTypes.h"
 
 namespace mondradiko {
 namespace assets {
-
-const size_t ASSET_REGISTRY_MAGIC_LENGTH = 4;
-const char ASSET_REGISTRY_MAGIC[] = "MDOA";
-
-const uint32_t ASSET_REGISTRY_MAX_LUMPS = 256;
-const uint32_t ASSET_LUMP_MAX_ASSETS = 4096;
-const uint32_t ASSET_LUMP_MAX_SIZE = 128 * 1024 * 1024;  // 128 MiB
-
-struct AssetRegistryHeader {
-  char magic[ASSET_REGISTRY_MAGIC_LENGTH];
-  uint32_t version;
-
-  uint32_t lump_count;
-};
-
-struct AssetRegistryLumpEntry {
-  LumpHash checksum;
-  LumpHashMethod hash_method;
-  LumpCompressionMethod compression_method;
-  uint32_t asset_count;
-};
-
-struct AssetRegistryEntry {
-  AssetId id;
-  uint32_t size;
-};
 
 class AssetBundle {
  public:
